blockifier: consolidate state-diff view trait, drop dead get_contract_addresses
Code review follow-up on the previous commit: move the state-diff view
trait next to CommitmentStateDiff (its natural home, and avoids a
module cycle with stateful_compression), give its accessors neutral
names shared by both impls instead of names borrowed from
CommitmentStateDiff's own fields, fold the repeated
storage/nonce/class-hash address chain into one default
`contract_addresses` method instead of duplicating it in two files, and
delete CommitmentStateDiff::get_contract_addresses now that both of its
callers were migrated to the trait.
