+++ crates/blockifier/src/state/stateful_compression.rs
@@ -6,8 +6,7 @@ use starknet_api::StarknetApiError;
 use starknet_types_core::felt::Felt;
 use thiserror::Error;
 
-use super::accessed_keys::StateDiffAccessedKeysView;
-use super::cached_state::{CachedState, StateMaps, StorageEntry};
+use super::cached_state::{CachedState, StateDiffView, StateMaps, StorageEntry};
 use super::errors::StateError;
 use super::state_api::{State, StateReader, StateResult};
 
@@ -85,22 +84,17 @@ pub fn allocate_aliases_in_storage<S: StateReader>(
 /// access pattern of `AliasUpdater::new` (counter read) plus each `AliasUpdater::insert_alias`
 /// call (one read per qualifying key).
 pub fn predicted_alias_storage_entries(
-    state_diff: &impl StateDiffAccessedKeysView,
+    state_diff: &impl StateDiffView,
     alias_contract_address: ContractAddress,
 ) -> HashSet<StorageEntry> {
     let mut entries = HashSet::new();
     entries.insert((alias_contract_address, ALIAS_COUNTER_STORAGE_KEY));
-    for address in state_diff
-        .storage_updates()
-        .keys()
-        .chain(state_diff.address_to_nonce_keys())
-        .chain(state_diff.address_to_class_hash().keys())
-    {
+    for address in state_diff.contract_addresses() {
         if should_compress_address(address) {
             entries.insert((alias_contract_address, StorageKey(address.0)));
         }
     }
-    for (address, inner) in state_diff.storage_updates() {
+    for (address, inner) in state_diff.storage_diffs() {
         for storage_key in inner.keys() {
             if should_compress_storage_key(storage_key, address) {
                 entries.insert((alias_contract_address, *storage_key));
